Extract `language-json`
