+++ src/language-js/index.js
@@ -1,9 +1,7 @@
 import estreePrinter from "./printer-estree.js";
-import estreeJsonPrinter from "./printer-estree-json.js";
 
 export const printers = {
   estree: estreePrinter,
-  "estree-json": estreeJsonPrinter,
 };
 export { default as languages } from "./languages.evaluate.js";
 export { default as options } from "./options.js";
