+++ src/language-json/index.js
@@ -1,0 +1,6 @@
+import estreeJsonPrinter from "./printer-estree-json.js";
+
+export const printers = {
+  "estree-json": estreeJsonPrinter,
+};
+export { default as languages } from "./languages.evaluate.js";
