+++ src/languages.js
@@ -5,3 +5,4 @@ export * as graphql from "./language-graphql/index.js";
 export * as markdown from "./language-markdown/index.js";
 export * as html from "./language-html/index.js";
 export * as yaml from "./language-yaml/index.js";
+export * as json from "./language-json/index.js";
